bot/requirements.txt
telethon==1.34.0
python-dotenv==1.0.0
pymongo==4.6.1
aiogram==3.3.0
requests==2.31.0 
